GPL of ASUS RT-N16 for firmware 9.0.0.4.380.2985

+++ release/src-rt/platform.mak
@@ -18,6 +18,13 @@ export HOST := arm-linux
 export TOOLS := $(SRCBASE)/toolchains/hndtools-arm-linux-2.6.36-uclibc-4.5.3
 export RTVER := 0.9.32.1
 export BCMSUB := brcmarm
+else ifeq ($(RTCONFIG_BCM_ARM_GCLIBC),y)
+export CROSS_COMPILE ?= arm-buildroot-linux-gnueabi-
+export CONFIGURE := ./configure arm-linux --build=$(BUILD)
+export TOOLCHAIN := $(shell cd $(dir $(shell which $(CROSS_COMPILE)gcc))/../.. && pwd -P)
+export CFLAGS += -fno-strict-aliasing
+SUBMAKE_SETTINGS += ARCH=$(ARCH)
+EXTRA_LDFLAGS := "-lgcc_s -lwlcsm"
 else
 ifeq ($(EXTRACFLAGS),)
 export EXTRACFLAGS := -DBCMWPA2 -fno-delete-null-pointer-checks -mips32 -mtune=mips32
@@ -57,6 +64,12 @@ define platformKernelConfig
 	else \
 		echo "# CONFIG_RGMII_BCM_FA is not set" >> $(1); \
 	fi; \
+	if [ "$(LACP)" = "y" ]; then \
+		sed -i "/CONFIG_LACP/d" $(1); \
+		echo "CONFIG_LACP=m" >>$(1); \
+		sed -i "/CONFIG_BCM_AGG/d" $(1); \
+		echo "CONFIG_BCM_AGG=y" >>$(1); \
+	fi; \
 	if [ "$(BCMNAND)" = "y" ]; then \
 		sed -i "/CONFIG_MTD_NFLASH/d" $(1); \
 		echo "CONFIG_MTD_NFLASH=y" >>$(1); \
@@ -77,34 +90,33 @@ define platformKernelConfig
 		echo "CONFIG_MTD_BRCMNAND=y" >>$(1); \
 	fi; \
 	if [ "$(ARM)" = "y" ]; then \
-		mkdir -p $(SRCBASE)/router/ctf_arm/linux;\
-		cp -f $(SRCBASE)/router/ctf_arm/bcm6x/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
+		mkdir -p $(SRCBASE)/router/ctf_arm/linux; \
 		if [ "$(BCM7)" = "y" ]; then \
 			if [ "$(ARMCPUSMP)" = "up" ]; then \
-				cp -f $(SRCBASE)/router/ctf_arm/bcm7_up/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
+				cp -f $(SRCBASE)/router/ctf_arm/bcm7_up/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
 			else \
-				cp -f $(SRCBASE)/router/ctf_arm/bcm7/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
-				cp -f $(SRCBASE)/router/dpsta/bcm7_3200/dpsta.o $(SRCBASE)/router/dpsta/linux;\
+				cp -f $(SRCBASE)/router/ctf_arm/bcm7/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
+				cp -f $(SRCBASE)/router/dpsta/bcm7_3200/dpsta.o $(SRCBASE)/router/dpsta/linux; \
 			fi; \
-		elif [ "$(BCM794)" = "y" ]; then \
-			if [ "$(ARMCPUSMP)" = "up" ]; then \
-				cp -f $(SRCBASE)/router/ctf_arm/bcm7_up/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
+		elif [ "$(BCM_7114)" = "y" ]; then \
+			if [ "$(GMAC3)" = "y" ]; then \
+				cp -f $(SRCBASE)/router/ctf_arm/bcm_7114_gmac3/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
 			else \
-				cp -f $(SRCBASE)/router/ctf_arm/bcm7/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
+				cp -f $(SRCBASE)/router/ctf_arm/bcm_7114/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
 			fi; \
 		elif [ "$(BCM10)" = "y" ]; then \
 			if [ "$(ARMCPUSMP)" = "up" ]; then \
-				cp -f $(SRCBASE)/router/ctf_arm/bcm7_up/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
+				cp -f $(SRCBASE)/router/ctf_arm/bcm7_up/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
 			else \
-				cp -f $(SRCBASE)/router/ctf_arm/bcm7/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
+				cp -f $(SRCBASE)/router/ctf_arm/bcm7/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
 			fi; \
+		elif [ "$(BCM9)" = "y" ]; then \
+                                cp -f $(SRCBASE)/router/ctf_arm/bcm9/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
 		else \
 			if [ "$(ARMCPUSMP)" = "up" ]; then \
-				cp -f $(SRCBASE)/router/ctf_arm/bcm6_up/linux/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
-				cp -f $(SRCBASE)/router/ufsd/broadcom_arm_up/ufsd.ko.46_up router/ufsd/broadcom_arm/ufsd.ko; \
-			fi; \
-			if [ "$(BWDPI)" = "y" ]; then \
-				cp -f $(SRCBASE)/router/ctf_arm/bcm6_iqos/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
+				cp -f $(SRCBASE)/router/ctf_arm/bcm6_up/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
+			else \
+				cp -f $(SRCBASE)/router/ctf_arm/bcm6/ctf.* $(SRCBASE)/router/ctf_arm/linux/; \
 			fi; \
 		fi; \
 	fi; \
@@ -240,28 +252,79 @@ define platformKernelConfig
 			fi; \
 			if [ -d $(SRCBASE)/router/wl_arm_7/prebuilt ]; then \
 				mkdir $(SRCBASE)/wl/linux ; \
-				cp $(SRCBASE)/router/wl_arm_7/prebuilt/wl_apsta.o $(SRCBASE)/wl/linux ; \
+				cp $(SRCBASE)/router/wl_arm_7/prebuilt/wl*.o $(SRCBASE)/wl/linux ; \
 				mkdir -p $(SRCBASE)/../../dhd/src/dhd/linux ; \
 				cp $(SRCBASE)/router/wl_arm_7/prebuilt/dhd.o $(SRCBASE)/../../dhd/src/dhd/linux ; \
 			fi; \
-		else \
-			if [ -d $(SRCBASE)/router/wl_arm/prebuilt ]; then \
+			if [ -d $(SRCBASE)/router/et_arm_7/prebuilt ]; then \
+				mkdir -p $(SRCBASE)/et/linux ; \
+				cp $(SRCBASE)/router/et_arm_7/prebuilt/et.o $(SRCBASE)/et/linux ; \
+			fi; \
+		elif [ "$(BCM_7114)" = "y" ]; then \
+			if [ -d $(SRCBASE)/router/wl_arm_7114/prebuilt ]; then \
+				mkdir -p $(SRCBASE)/../dhd/src/dhd/linux ; \
+				cp $(SRCBASE)/router/wl_arm_7114/prebuilt/dhd.o $(SRCBASE)/../dhd/src/dhd/linux ; \
+			fi; \
+			if [ -d $(SRCBASE)/router/et_arm_7114/prebuilt ]; then \
+				mkdir -p $(SRCBASE)/et/linux ; \
+				cp $(SRCBASE)/router/et_arm_7114/prebuilt/et.o $(SRCBASE)/et/linux ; \
+			fi; \
+		elif [ "$(BCM_10)" = "y" ]; then \
+			if [ -d $(SRCBASE)/router/wl_arm_10/prebuilt ]; then \
 				mkdir $(SRCBASE)/wl/linux ; \
-				cp $(SRCBASE)/router/wl_arm/prebuilt/wl_apsta.o $(SRCBASE)/wl/linux ; \
-			elif [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME) ]; then \
+				cp $(SRCBASE)/router/wl_arm_10/prebuilt/wl*.o $(SRCBASE)/wl/linux ; \
+			fi; \
+			if [ -d $(SRCBASE)/router/et_arm_10/prebuilt ]; then \
+				mkdir -p $(SRCBASE)/et/linux ; \
+				cp $(SRCBASE)/router/et_arm_10/prebuilt/et.o $(SRCBASE)/et/linux ; \
+			fi; \
+		elif [ "$(BCM9)" = "y" ]; then \
+			if [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME) ]; then \
 				if [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/linux ]; then \
 					cp -rf $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/linux $(SRCBASE)/wl/. ; \
 				fi; \
 				if [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/clm ]; then \
 					cp -f $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/clm/src/wlc_clm_data.c $(SRCBASE)/wl/clm/src/. ; \
 				fi; \
-			else \
+			elif [ -d $(SRCBASE)/wl/sysdeps/default ]; then \
 				if [ -d $(SRCBASE)/wl/sysdeps/default/linux ]; then \
 					cp -rf $(SRCBASE)/wl/sysdeps/default/linux $(SRCBASE)/wl/. ; \
 				fi; \
 				if [ -d $(SRCBASE)/wl/sysdeps/default/clm ]; then \
 					cp -f $(SRCBASE)/wl/sysdeps/default/clm/src/wlc_clm_data.c $(SRCBASE)/wl/clm/src/. ; \
 				fi; \
+			fi; \
+			if [ -d $(SRCBASE)/router/wl_arm_9/prebuilt ]; then \
+				mkdir $(SRCBASE)/wl/linux ; \
+				cp $(SRCBASE)/router/wl_arm_9/prebuilt/wl*.o $(SRCBASE)/wl/linux ; \
+			fi; \
+			if [ -d $(SRCBASE)/router/et_arm_9/prebuilt ]; then \
+				mkdir -p $(SRCBASE)/et/linux ; \
+				cp $(SRCBASE)/router/et_arm_9/prebuilt/et.o $(SRCBASE)/et/linux ; \
+			fi; \
+		else \
+			if [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME) ]; then \
+				if [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/linux ]; then \
+					cp -rf $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/linux $(SRCBASE)/wl/. ; \
+				fi; \
+				if [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/clm ]; then \
+					cp -f $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/clm/src/wlc_clm_data.c $(SRCBASE)/wl/clm/src/. ; \
+				fi; \
+			elif [ -d $(SRCBASE)/wl/sysdeps/default ]; then \
+				if [ -d $(SRCBASE)/wl/sysdeps/default/linux ]; then \
+					cp -rf $(SRCBASE)/wl/sysdeps/default/linux $(SRCBASE)/wl/. ; \
+				fi; \
+				if [ -d $(SRCBASE)/wl/sysdeps/default/clm ]; then \
+					cp -f $(SRCBASE)/wl/sysdeps/default/clm/src/wlc_clm_data.c $(SRCBASE)/wl/clm/src/. ; \
+				fi; \
+			fi; \
+			if [ -d $(SRCBASE)/router/wl_arm/prebuilt ]; then \
+				mkdir $(SRCBASE)/wl/linux ; \
+				cp $(SRCBASE)/router/wl_arm/prebuilt/wl*.o $(SRCBASE)/wl/linux ; \
+			fi; \
+			if [ -d $(SRCBASE)/router/et_arm/prebuilt ]; then \
+				mkdir -p $(SRCBASE)/et/linux ; \
+				cp $(SRCBASE)/router/et_arm/prebuilt/et.o $(SRCBASE)/et/linux ; \
 			fi; \
 		fi; \
 	else \
